django_api/requirements.txt
psycopg2-binary==2.9
django==4.0.4
flake8==4.0
python-dotenv==0.21.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-split-settings==1.2.0
django-admin-rangefilter==0.9.0
django-debug-toolbar==3.4.0
isort==5.10.1
typing==3.7.4.3
pytest==7.1.3
uwsgi==2.0.20
